// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Network = {
  TESTNET: 'TESTNET',
  PUBLIC: 'PUBLIC'
} as const

export type Network = (typeof Network)[keyof typeof Network]


export const WalletStatus = {
  ACTIVE: 'ACTIVE',
  PAUSED: 'PAUSED'
} as const

export type WalletStatus = (typeof WalletStatus)[keyof typeof WalletStatus]


export const WebhookStatus = {
  ACTIVE: 'ACTIVE',
  PAUSED: 'PAUSED'
} as const

export type WebhookStatus = (typeof WebhookStatus)[keyof typeof WebhookStatus]


export const DeliveryStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  EXHAUSTED: 'EXHAUSTED'
} as const

export type DeliveryStatus = (typeof DeliveryStatus)[keyof typeof DeliveryStatus]


export const Plan = {
  FREE: 'FREE',
  PRO: 'PRO',
  SCALE: 'SCALE'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]


export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PriceType = {
  FIXED: 'FIXED',
  CUSTOM: 'CUSTOM'
} as const

export type PriceType = (typeof PriceType)[keyof typeof PriceType]


export const PaymentStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]